requirements.txt
scrapy==2.11.2
playwright==1.44.0
supabase==2.4.6
python-dotenv==1.0.1
pandas==2.2.2
geopy==2.4.1
folium==0.17.0
streamlit==1.35.0
streamlit-folium==0.21.1
anthropic==0.28.0
fake-useragent==1.5.1
requests==2.32.3
python-telegram-bot==21.3
